refactor: consolidate access control middlewares
  - Merged `adminMiddleware.ts`, `workerMiddleware.ts`, and `clientMiddleware.ts` into a unified `accessMiddleware.ts` file to centralize role-based access logic.
  - Updated import paths across all routes (`chat.ts`, `dashboard.ts`, `orders.ts`, `proposals.ts`, `reports.ts`, `specializations.ts`, and financial routes) to point to the unified `accessMiddleware.ts`.
  - Enhanced `/orders` routes with granular role checks: added `authorizeClient` to creation, cancellation, and rating endpoints; added `authorizeWorker` to start and finish work endpoints.
  - Cleaned up `/proposals` routes by removing conflicting chained middlewares (e.g., `authorizeClient` followed by `authorizeWorker`) which prevented valid access.
  - Updated `CHANGELOG.md` with the latest changes.
