+++ src/routes/v1/orders.ts
@@ -31,6 +31,7 @@ import {
 } from '../../controllers/NegotiationController.js';
 import { CreateNegotiationSchema } from '../../schemas/requests/negotiation.request.js';
 import { parseFormDataJson } from 'src/middlewares/multiformParserMiddleware.js';
+import { authorizeClient, authorizeWorker } from 'src/middlewares/accessMiddleware.js';
 
 
 const router = Router();
@@ -42,20 +43,28 @@ router.post(
   upload.array('images', 3),
   parseFormDataJson('orderData'),
   validateBody(CreateOrderSchema),
+  authorizeClient,
   orderController.create
 );
 router.get('/', validateQuery(OrderQuerySchema), orderController.list);
 router.get('/:orderId', validateParams(OrderIdParamsSchema), orderController.getById);
-router.delete('/:orderId', validateParams(OrderIdParamsSchema), orderController.cancel);
+router.delete('/:orderId', validateParams(OrderIdParamsSchema), authorizeClient, orderController.cancel);
 router.get('/:orderId/location', validateParams(OrderIdParamsSchema), orderController.getLocation);
-router.post('/:orderId/start-work', validateParams(OrderIdParamsSchema), orderController.startWork);
+router.post(
+  '/:orderId/start-work',
+  validateParams(OrderIdParamsSchema),
+  authorizeWorker,
+  orderController.startWork
+);
 router.post(
   '/:orderId/finish-work',
+  authorizeWorker,
   validateParams(OrderIdParamsSchema),
   orderController.finishWork
 );
 router.post(
   '/:orderId/rate',
+  authorizeClient,
   validateParams(OrderIdParamsSchema),
   validateBody(OrderRateSchema),
   orderController.rate
